// src/Api/models/cart-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { Product } from "./product";

/**
 *
 * @export
 * @interface CartItem
 */
export interface CartItem {
  /**
   *
   * @type {number}
   * @memberof CartItem
   */
  id?: number;
  /**
   *
   * @type {string}
   * @memberof CartItem
   */
  size?: string;
  /**
   *
   * @type {number}
   * @memberof CartItem
   */
  quantity?: number;
  /**
   *
   * @type {number}
   * @memberof CartItem
   */
  maxPrice?: number;
  /**
   *
   * @type {number}
   * @memberof CartItem
   */
  sellingPrice?: number;
  /**
   *
   * @type {number}
   * @memberof CartItem
   */
  userId?: number;
  /**
   *
   * @type {Product}
   * @memberof CartItem
   */
  product?: Product;
}